+++ WidgetInstaller.ps1
@@ -29,6 +29,7 @@ $script:WidgetRuntimeFiles = @(
     'ModelRequestRecorder.ps1'
     'UsageHistory.ps1'
     'UsageReport.ps1'
+    'WidgetTrend.ps1'
     'WidgetConfig.ps1'
     'WidgetStrings.ps1'
     'WidgetInstaller.ps1'